refactor: HandlerMapping에 init 메소드 추상화
