+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -20,7 +20,8 @@ public class ManualHandlerMapping implements HandlerMapping {
 
     private static final Map<String, Controller> controllers = new HashMap<>();
 
-    public void initialize() {
+    @Override
+    public void init() {
         controllers.put("/", new ForwardController("/index.jsp"));
         controllers.put("/login", new LoginController());
         controllers.put("/login/view", new LoginViewController());
