+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/mapping/HandlerMapping.java
@@ -3,5 +3,7 @@ package com.interface21.webmvc.servlet.mvc.handler.mapping;
 import jakarta.servlet.http.HttpServletRequest;
 
 public interface HandlerMapping {
+    void init();
+
     Object getHandler(HttpServletRequest request);
 }
